tests/wpt/web-platform-tests/css/css-scroll-snap/snap-to-visible-areas.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
div {
  position: absolute;
  margin: 0px;
}
#scroller {
  height: 600px;
  width: 600px;
  overflow: scroll;
  scroll-snap-type: both mandatory;
}
#space {
  width: 2000px;
  height: 2000px;
}

.snap {
  width: 200px;
  height: 200px;
  background-color: blue;
  scroll-snap-align: start;
}

#left-top {
  left: 0px;
  top: 0px;
}

#right-top {
  left: 800px;
  top: 0px;
}

#left-bottom {
  left: 0px;
  top: 800px;
}

#right-bottom {
  left: 1800px;
  top: 1800px;
  scroll-margin-top: 1000px;
  scroll-margin-left: 1000px;
}

</style>
<div id="scroller">
  <div id="space"></div>
  <div id="left-top" class="snap"></div>
  <div id="right-top" class="snap"></div>
  <div id="left-bottom" class="snap"></div>
  <div id="right-bottom" class="snap"></div>
</div>
<script>
var scroller = document.getElementById("scroller");
test(() => {
  scroller.scrollTo(0, 0);
  assert_equals(scroller.scrollLeft, 0);
  assert_equals(scroller.scrollTop, 0);
  scroller.scrollTo(300, 0);
  assert_equals(scroller.scrollLeft, 800);
  assert_equals(scroller.scrollTop, 0);
}, 'Only snap to visible area on X axis, even when the non-visible ones are closer');

test(() => {
  scroller.scrollTo(0, 0);
  assert_equals(scroller.scrollLeft, 0);
  assert_equals(scroller.scrollTop, 0);
  scroller.scrollTo(0, 300);
  assert_equals(scroller.scrollLeft, 0);
  assert_equals(scroller.scrollTop, 800);
}, 'Only snap to visible area on Y axis, even when the non-visible ones are closer');

test(() => {
  scroller.scrollTo(0, 0);
  assert_equals(scroller.scrollLeft, 0);
  assert_equals(scroller.scrollTop, 0);
  scroller.scrollTo(300, 300);
  assert_equals(scroller.scrollLeft, 800);
  assert_equals(scroller.scrollTop, 800);
}, 'snap to snap area inflated by scroll-margin, even when the non-visible ones are closer');
</script>
